фиксанул сундуки, поменял структуру малясь
