+++ entities/living_entities.py
@@ -1,11 +1,12 @@
 from random import randint
-from typing import List
-from entities.item import ItemEntity
-from items import Item
+from typing import TYPE_CHECKING, List
 from pygame_entities.entities.entity import Entity
 from pygame_entities.utils.math import Vector2, clamp
-
 from entities.map import Map
+
+if TYPE_CHECKING:
+    from items import Item
+    from entities.item import ItemEntity
 
 
 class LivingEntity(Entity):
@@ -13,9 +14,8 @@ class LivingEntity(Entity):
     Живая сущность. У нее есть здоровье (HP), которое можно изменять. Так же она может умереть, если HP будут равны 0.
     """
 
-    def __init__(self, position: Vector2, max_hp: int, map: Map) -> None:
+    def __init__(self, position: Vector2, max_hp: int) -> None:
         super().__init__(position)
-        self.map = map
         self.max_hp = max_hp
         self.hp = max_hp
 
@@ -40,7 +40,7 @@ class LivingEntity(Entity):
         self.add_hp(amount)
         self.hp = clamp(self.hp, 0, self.max_hp)
 
-    def get_loot(self) -> List[Item]:
+    def get_loot(self) -> List["Item"]:
         """
         тут будет генерироваться и возвращаться список предметов, которые должны выпасть из этой сущности при смерти.
         """
