+++ items/__init__.py
@@ -1,0 +1,4 @@
+from items.inventory import Inventory
+from items.items import Iron, IronIngot, Gold, Wood, GoldIngot
+from items.recipes import BaseRecipe
+from items.item import Item, UsableItem
