+++ items/recipes.py
@@ -1,9 +1,7 @@
-import dataclasses
 from typing import List, Tuple
-from items import Item
+from items.item import Item
 
 
-@dataclasses
 class BaseRecipe:
     """Класс для рецептика"""
     ingredients: List[Tuple[Item, int]]
